[MOVE/#151] 프로필 및 계정관리 패키지 분리
